pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "repo-graphrag-mcp"
version = "0.2.0"
description = "GraphRAG-based MCP server: builds a repository knowledge graph (Tree-sitter + LightRAG) for Q&A and change planning."
readme = "README.md"
requires-python = ">=3.10"
license = { text = "MIT" }

dependencies = [
    "lightrag-hku==1.4.7",
    "anthropic==0.66.0",
    "openai>=1.0.0",
    "google-genai>=1.31.0",
    "transformers==4.56.1",
    "tree-sitter==0.25.0",
    "tree-sitter-python==0.23.6",
    "tree-sitter-cpp==0.23.4",
    "tree-sitter-java==0.23.5",
    "tree-sitter-javascript==0.23.1",
    "tree-sitter-typescript==0.23.2",
    "tree-sitter-rust==0.24.0",
    "tree-sitter-c==0.24.1",
    "tree-sitter-c-sharp>=0.23.1",
    "tree-sitter-go==0.23.4",
    "tree-sitter-ruby>=0.23.1",
    "tree-sitter-html>=0.23.2",
    "tree-sitter-css>=0.23.2",
    "python-dotenv>=1.0.0",
    "fastmcp>=0.1.0",
    "torch==2.8.0",
    "numpy==2.2.6",
    "tokenizers>=0.21.4",
    "sentence-transformers>=3.0.1",
    "faiss-cpu>=1.7.0"
]

[project.optional-dependencies]
dev = [
    "pytest>=8.4.2",
]

[project.scripts]
repo-graphrag-mcp = "server:mcp.run"

[tool.hatch.build.targets.wheel]
packages = ["repo_graphrag"]
include = [
    "server.py",
    "repo_graphrag/**/*",
]


